Adds tests
