+++ src/Components/Publishing/Display/Canvas/index.tsx
@@ -3,6 +3,7 @@ import React from "react"
 import Waypoint from "react-waypoint"
 import styled, { StyledFunction } from "styled-components"
 import Colors from "../../../../Assets/Colors"
+import Events from "../../../../Utils/Events"
 import { track } from "../../../../Utils/track"
 import { pMedia } from "../../../Helpers"
 import { Fonts } from "../../Fonts"
