+++ src/Components/Publishing/Display/track-once.ts
@@ -17,7 +17,7 @@ const alreadyFired = {}
  *        }
  *      }
  */
-export function trackImpression(unitLayout) {
+const trackOnce = (unitLayout, action) => {
   return (target, name, descriptor) => {
     const decoratedFn = descriptor.value
     // tslint:disable-next-line:only-arrow-functions
@@ -25,12 +25,13 @@ export function trackImpression(unitLayout) {
       const key = [
         this.props.campaign.name,
         unitLayout(this.props),
+        action,
         (this.props.article && this.props.article.id) || this._reactInternalInstance._debugID,
       ].join(":")
       if (alreadyFired[key]) return decoratedFn.apply(this, arguments)
       this.props.tracking &&
         this.props.tracking.trackEvent({
-          action: "Impression",
+          action,
           entity_type: "display_ad",
           campaign_name: this.props.campaign.name,
           unit_layout: unitLayout(this.props),
@@ -41,26 +42,10 @@ export function trackImpression(unitLayout) {
   }
 }
 
+export function trackImpression(unitLayout) {
+  return trackOnce(unitLayout, "Impression")
+}
+
 export function trackViewability(unitLayout) {
-  return (target, name, descriptor) => {
-    const decoratedFn = descriptor.value
-    // tslint:disable-next-line:only-arrow-functions
-    descriptor.value = function () {
-      const key = [
-        this.props.campaign.name,
-        unitLayout(this.props),
-        (this.props.article && this.props.article.id) || this._reactInternalInstance._debugID,
-      ].join(":")
-      if (alreadyFired[key]) return decoratedFn.apply(this, arguments)
-      this.props.tracking &&
-        this.props.tracking.trackEvent({
-          action: "Viewability",
-          entity_type: "display_ad",
-          campaign_name: this.props.campaign.name,
-          unit_layout: unitLayout(this.props),
-        })
-      alreadyFired[key] = true
-      decoratedFn.apply(this, arguments)
-    }
-  }
+  return trackOnce(unitLayout, "Viewability")
 }
